add envvars it to both SessionStartRequest and LaunchRequest
